[snippy] Fix access alignment for burst mode
Previously the code did the unthinkable:

```
StrongestAR.AccessAlignment =
    OpcodeToAR.at(*StrongestAccessSize).AccessAlignment;
```

That is most certainly broken. The correct thing to do is to
get the maximum required alignment, not the alignment of
the largest access size. This would lead to unintentional
generation of misaligned accesses when the same base register
is used for the case like:

- `AccessSize` = `4`, `AccessAlignment` = `4`
- `AccessSize` = `8`, `AccessAlignment` = `1`

In this case the minimum required alignment is `4`, but the
code prior to this change would drop the restriction to `1`.

While we are at it, also clean up this copy-paste pasta with some macros.