[llvm-ie] aggregate several minor improvements (#309)
* This patch consolidates following enhancements:

* [llvm-ie] enable vendor extensions
Enables llvm-ie to work with RISC-V vendor extensions. Note
that instructions from a particular vendor extension will not appear in
the output unless the corresponding extension is explicitly specified.
You can think of it as vendor extensions having lower significance than
standard ones, so you are guaranteed that vendor instructions won't
appear unless a vendor extension is specified.

* [llvm-ie] add llvm-ie documentation

* [llvm-ie] make pseudo instructions available for llvm-ie
Makes the llvm-ie tool work with pseudo instructions by
default. Additionally, it introduces a new option -disable-pseudo that
removes pseudo instructions from the output, therefore preserving the
old behavior.

* [llvm-ie] add verbose output option
This patch adds a `--verbose` (`-v`) option to the `llvm-ie` tool. When
specified, the tool's output is augmented with instruction categories
for each instruction.
Currently, an instruction's categories represent the set of RISC-V
extensions it belongs to. Specifying any extension from this set will
cause the corresponding instruction to appear in the output. This option
is useful for understanding why a particular instruction is included in
the results.

* [llvm-ie] fix extension.test (#310)

---------

Co-authored-by: Daniil Avdeev <daniil.avdeev@syntacore.com>